+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Platform Version 0.9.17 - UNRELEASED
 * Change default values of TopicProducerConfig ([#2069](https://github.com/infinyon/fluvio/issues/2069))
+* Enhance CLI Diagnostics with system info  ([#2069](https://github.com/infinyon/fluvio/pull/2092))
 
 ## Platform Version 0.9.16 - 2022-01-03
 * Consume with end ([#1940](https://github.com/infinyon/fluvio/issues/1940))